+++ spacepackets/cfdp/pdu/file_data.py
@@ -5,7 +5,7 @@ from typing import Union, Optional
 import struct
 
 from spacepackets.cfdp import LargeFileFlag, CrcFlag
-from spacepackets.cfdp.defs import Direction
+from spacepackets.cfdp.defs import Direction, TransmissionMode
 from spacepackets.cfdp.pdu.file_directive import SegmentMetadataFlag, PduType
 from spacepackets.cfdp.conf import PduConfig
 from spacepackets.cfdp.pdu.header import PduHeader, AbstractPduBase
@@ -90,6 +90,10 @@ class FileDataPdu(AbstractPduBase):
     @property
     def pdu_type(self) -> PduType:
         return self.pdu_header.pdu_type
+
+    @property
+    def transmission_mode(self) -> TransmissionMode:
+        return self.pdu_header.transmission_mode
 
     @property
     def direction(self) -> Direction: